so many changes

+++ src/main/java/net/rom/exoplanets/astronomy/yzcetisystem/b/WorldProviderYzCetiB.java
@@ -56,7 +56,7 @@ public class WorldProviderYzCetiB extends WorldProviderExoPlanet {
     @Override
     public float getPlanetTemp() {
         ExoPlanet planet = this.getPlanet();
-        float planetTemp = planet.getPlanetTemp();
+        float planetTemp = (float) planet.getPlanetTemp();
         if (!this.isDaytime()) {
             planetTemp /= 1.4F;
         }
